Flask_Project/Flask_Project/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Cors
psycopg2-binary
